SOF-1743 'SystemKeybindings' are only updated when there is a Rundown change. 'ActionTriggerKeybindings' are only updated when there is an update to ActionTriggers or Actions
