+++ src/app/settings/keyboard-mapping-settings/components/edit-keyboard-mapping/edit-keyboard-mapping.component.ts
@@ -33,7 +33,7 @@ export class EditKeyboardMappingComponent implements OnInit {
       label: [this.keyboardMapping?.actionTrigger.data.label ?? ''],
       keys: [this.keyboardMapping?.actionTrigger.data.keys, [Validators.required]],
       triggerOn: [KeyEventType.RELEASED, [Validators.required]],
-      mappedToKeys: [this.keyboardMapping?.actionTrigger.data.mappedToKeys],
+      mappedToKeys: [this.keyboardMapping?.actionTrigger.data.mappedToKeys ?? []],
     })
 
     this.actionTriggerForm = this.formBuilder.group({
